# Optimize host merging and handling