This PR optimizes the processes of handling and merging hosts. By consolidating most DB transactions and performing bulk reads, handling hosts is up to 90x faster, while merging hosts is up to 60x faster.

## Handling collected hosts

Handling of source hosts has been optimized by fetching all hostnames in a single DB query and performing all work in a single transaction as opposed to one transaction per host.

### Results

When handling >13k hosts from 5 different sources, the following results were observed:

**Before**

```
[root] Done handling hosts from source, 'nivlheim', in 289.80 seconds. 
[root] Done handling hosts from source, 'mreg', in 104.38 seconds.
[root] Done handling hosts from source, 'nav', in 2.48 seconds.
[root] Done handling hosts from source, 'nav-ea', in 20.37 seconds.
```

**After**

```
[root] Done handling hosts from source, 'nivlheim', in 3.58 seconds.
[root] Done handling hosts from source, 'mreg', in 1.76 seconds.
[root] Done handling hosts from source, 'nav', in 0.17 seconds. 
[root] Done handling hosts from source, 'nav-ea', in 0.23 seconds
```

**Summary**

~ 15-90x faster

## Merging hosts

Merging of hosts has similarly been sped up by consolidating transactions and performing bulk reads. 

### Results

Merging the >13k hosts collected from the sources into >10k hosts took the following time:

**Before**

```
[root] Done with merge in 464.68 seconds. Equal hosts: 10115
[root] Done with merge in 464.78 seconds. Equal hosts: 10111
[root] Done with merge in 456.70 seconds. Equal hosts: 10111
[root] Done with merge in 458.27 seconds. Equal hosts: 10111
```

**After**

```
[root] Done with merge in 9.82 seconds. Equal hosts: 10039
[root] Done with merge in 7.75 seconds. Equal hosts: 10041
[root] Done with merge in 7.60 seconds. Equal hosts: 10041
[root] Done with merge in 7.51 seconds. Equal hosts: 10041
[root] Done with merge in 7.41 seconds. Equal hosts: 10041
```

**Summary**

~ 50-60x faster