@pederhan 2023-08-08
W have tested this in production for some time now, and no problems have been observed.